feat(sound): add daiku command (Ode to Joy PSG demo)
- Add kernel/daiku.c: plays Beethoven's Ode to Joy (public domain)
  BPM=120, 4/4 time, 8 bars, melody on ch0 + bass on ch1, 40ms gap
- Replace furusato command with daiku in shell.c
  Old: furusato (do_fork + manual reparent to PID1)
  New: daiku (ring-3 double-fork pattern via fork()+exec_fn())
  The launcher child fork()s a grandchild, exec_fn(daiku_main),
  then immediately exit()s; the grandchild is reparented to PID1
  by exit.c reparent logic and plays in the background
- Remove kernel/furusato.c (no longer referenced)"
